pyproject.toml
[project]
name = "airdec-backend"
version = "0.1.0"
description = "backend for airdec ai extraction"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "typer>=0.15.0",
    "fastapi-limiter>=0.2.0",
    "fastapi[standard]>=0.112.1",
    "httpx>=0.27.0",
    "nanoid>=2.0.0",
    "pdfplumber>=0.11.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.0.0",
    "pydantic-ai[openai]>=0.0.49",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.37",
    "temporalio>=1.7.0",
    "PyJWT>=2.0.0",
    "cryptography>=44.0.0",
]

[project.scripts]
airdec = "app.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
]
ignore = [
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
    "D106",  # missing docstring in public nested class
]

[tool.ruff.lint.pydocstyle]
convention = "google"
